회원 탈퇴 로직 완료

+++ src/pages/mypage/SignOut.jsx
@@ -9,12 +9,11 @@ const SignOut = () => {
   const [user_name, setUser_name] = useState("");
   const [user_jumin, setUser_jumin] = useState("");
   const [member, setMember] = useState([]);
-  const [showModal, setShowModal] = useState(false);
-  const [modalMessage, setModalMessage] = useState("");
+  const [isMatching, setIsMatching] = useState(false);
 
   useEffect(() => {
     const membersInfo = async () => {
-      // // 로컬스토리지에서 로그인한 사용자 정보 가져 오기
+      // 로컬스토리지에서 로그인한 사용자 정보 가져오기
       const loginUserEmail = sessionStorage.getItem("user_id");
       try {
         const rsp = await AxiosApi.memberSelect(loginUserEmail); // 회원 정보 가져오기
@@ -28,14 +27,36 @@ const SignOut = () => {
     membersInfo();
   }, []);
 
+  useEffect(() => {
+    const sessionUserName = sessionStorage.getItem("user_name");
+    const sessionUserJumin = sessionStorage.getItem("user_jumin");
+
+    if (user_name === sessionUserName && user_jumin === sessionUserJumin) {
+      setIsMatching(true);
+    } else {
+      setIsMatching(false);
+    }
+  }, [user_name, user_jumin]);
+
   const memberDelete = async () => {
     try {
-      const isMemberValid = await AxiosApi.memberCheck(user_name, user_jumin);
-      if (isMemberValid) {
-        const rsp = await AxiosApi.memberDelete(user_name, user_jumin);
-        sessionStorage.clear();
+      const sessionUserName = sessionStorage.getItem("user_name");
+      const sessionUserJumin = sessionStorage.getItem("user_jumin");
+
+      if (user_name === sessionUserName && user_jumin === sessionUserJumin) {
+        const isMemberValid = await AxiosApi.memberCheck(user_name, user_jumin);
+        if (isMemberValid) {
+          await AxiosApi.memberDelete(user_name, user_jumin);
+          sessionStorage.clear();
+          alert("회원 탈퇴 되었습니다.");
+          navigate(`/`);
+        } else {
+          alert("회원 정보를 확인하는 데 실패했습니다.");
+        }
+      } else {
+        alert("이름과 주민등록번호가 일치하지 않습니다.");
+        setIsMatching(false);
       }
-      navigate(`/`);
     } catch (e) {
       console.log(e);
     }
@@ -50,19 +71,26 @@ const SignOut = () => {
         <input
           type="text"
           placeholder="이름"
+          // value={user_name}
           onChange={(e) => setUser_name(e.target.value)}
         />
         <input
           type="text"
           placeholder="주민등록번호"
+          // value={user_jumin}
           onChange={(e) => setUser_jumin(e.target.value)}
         />
         <p></p>
-        <div className={styles.finalCheck} onClick={memberDelete}>
+        <div
+          className={styles.finalCheck}
+          onClick={memberDelete}
+          // style={{ cursor: isMatching ? "pointer" : "not-allowed" }}
+        >
           탈퇴
         </div>
       </div>
     </div>
   );
 };
+
 export default SignOut;
